requirements.txt
arxiv==2.1.0
beautifulsoup4==4.12.3
cohere==5.5.0
duckduckgo_search==5.3.0b4
fitz==0.0.1.dev2
g4f==0.3.1.0
google_search_results==2.4.2
googlesearch_python==1.2.3
groq==0.8.0
langchain==0.2.0
langchain_community==0.2.0
langchain_core==0.2.1
langchain_openai==0.1.7
langchain_text_splitters==0.2.0
matplotlib==3.8.2
ollama==0.2.0
openai==1.30.2
playwright==1.43.0
PyGithub==2.3.0
pypdf==4.2.0
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.34.0
tenacity==8.3.0
termcolor==2.4.0
together==1.2.0
urllib3==2.2.1
youtube_transcript_api==0.6.2
yt_dlp==2023.12.30